Interactive version & alignment improvements
DB.[ch], QV.[ch], and align.[ch] can now be compiled in an interactive
version by setting the defined constant INTERACTIVE in DB.h.  Rather
than printing an error message to stderr and exiting with code 1, the
interactive version of the routines place an error message in a buffer
EPLACE and return an error code to the caller.

The aligner used O(N+M) space where N and M are the lengths of the two
“reads” being compared.  Now a fixed amount of memory is used so that
one can compare for example 2 chromosomes against each other without
exceeding memory.

The daligner filter has been upgraded so that when comparing two very
large sequences (say entire chromosomes), it is guaranteed to explore
and report all independent alignments paths in the underlying edit
graph.