Daligner upgrade, LAdump + LAindex utilities added.  New developer alignment routines.
daligner’s overlapping alignment artifact has been eliminated.  It also
manages space better allowing it to be more sensitive for a given
amount of memory.

A new LAdump utility has been added that allows one to list any desired
information about one or more A-piles in a .las file.  The output
format is designed to make reading in by a subsequent program as easy
as possible.

A new LAindex utility has been added that for a given .las file, builds
a .las.idx file that contains the offset in the .las file to the start
of each A-pile.  Not currently used by any of our programs, provided
for developers.

A bug in Lashow when comparing sequences against themselves has been
fixed.

LAsort works even when overlaps with different B orientations are in
the file.

For developers, the alignment suite has been extended so that the trace
routines can compute lowermost, uppermost, or greediest optimal
alignment between every pair of trace points.  There is also a new
routine Find_Extension that attempts to extend a given alignment left
or right from a given point in an edit matrix.