DBdust 0-div bug fix, daligner works with k=32
The daligner has been upgraded so that it will work when k=32 (and
reports an error for k>32).  Another small problem when a track was
associated with the source block has also been fixed.

As per Chris Dunn’s push, a divide by zero with empty tracks was fixed
and DBstat’s code was further improved to check for divide by zero if
the DB itself is empty.