Enhanced HPC script generators & miscellany
HPC.daligner and HPC.mapper produce scripts that are hopefully a little
easier to use, including separate batch commands for removing and
checking .las files.

The DB library, DB.[ch] + QV.[ch], has a new routine that returns the
size in bytes of a currently loaded database block.

LAcheck now returns a non-zero status if it exits on an error.

A few minor edits to shut down warnings during compiles.