HPCmapper & HPCdaligner merged into HPC.daligner
These two programs have been merged into one that generates a
daligner script if supplied one DB as an argument, and a mapper script
if supplied two.  This change is being made in part to make room for
the DAMAPPER module where in a customizer version of daligner maps
reads and it include HPC.damapper which is uncomfortably similar to
HPC.mapper.  I have also added -a sorting options to LAsort and LAmerge
in order to properly sort .las files corresponding to read mappings.