LAshow display modifications
LAshow now gives the B-interval coordinates as location in the
(uncomplemented) B sequence from high to low.  This is easier for human
beings to understand as it tells you where the interval is in B (which
is much more informative then knowing where it is in the complement of
B (the encoding used by us internally)).