Upgrade to HPC.aligner and other small changes
The routine Local_Alignment of the core align.[ch] sub-library now
considers and handles the case where the A-sequence is complemented (as
opposed to the B-sequence).  This is indicated with a new flag
ACOMP_FLAG in the Overlap record (see align.h).

A bug was fixed in the DB sub-library involving loading block tracks.

Handle redundancies of filter.c of the aligner has been improved to
truly removal all overlapping local alignments.  Some “duplicates” were
getting through and the main trick was to consider alignments with the
same start and end points as being “Entwined”.

HPC.aligner now has -f and -d options to make large job execution
easier and more efficient of OS resources.

Cosmetic changes to LAmerge and LAshow, and the README is updated to
reflect new features.