-T option added and all LA programs are now chain-aware
-T option to daligner and HPC.daligner allow one to specify number of
threads to use to a power of 2.

The damapper outputs *chains* of LAs and encodes this in the “flags”
field of Overlap and Alignment records.  LAsort and LAmerge are now
chain aware and when they encounter files with the flags, they sort
chains as a unit, based on their first (smallest index) LA.  Similarly
LAshow and LAdump display the chain information, and LAcheck checks
chains and flag settings for validity.